Starting to implement the creation wizard
