+++ src/App.tsx
@@ -5,7 +5,7 @@ import { Route, Switch } from "react-router"
 import Layout from "./components/Layout"
 import Background from "./components/Background"
 import Home from "./components/Home"
-import CreateDao from "./components/DaoCreation/CreateDao"
+import SummaryView from "./components/DaoCreator/SummaryView"
 import Founders from "./components/Founders"
 import Notifications from "./components/Notifications"
 
@@ -14,8 +14,8 @@ const App: SFC = () => (
     <Background />
     <Switch>
       <Route exact path="/" component={Home} />
-      <Route exact path="/create-dao" component={CreateDao} />
-      <Route exact path="/configure-founders" component={Founders} />
+      <Route path="/create-dao/summary" component={SummaryView} />
+      <Route path="/create-dao/founders" component={Founders} />
     </Switch>
     <Notifications />
   </Layout>
