+++ src/components/DaoCreator/SummaryView/index.ts
@@ -1,0 +1,3 @@
+import Summary from "./summary.container"
+
+export default Summary
